feat(converter): create separate converter project
