+++ AddIn/EPFL.SpeckleTopSolid.AddIn.csproj
@@ -18,7 +18,7 @@
     <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x64'">
-    <OutputPath>..\..\..\..\..\..\Sources\Topsolid 7.15\Debug x64\</OutputPath>
+    <OutputPath>C:\Sources\Topsolid 7.15\Debug x64\</OutputPath>
     <PlatformTarget>x64</PlatformTarget>
     <NoWarn>1607</NoWarn>
     <DefineConstants>DEBUG;TRACE</DefineConstants>
