+++ EPFL.SpeckleTopSolid.AddIn.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EPFL.SpeckleTopSolid.AddIn"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EPFL.SpeckleTopSolid.UI", "UI\EPFL.SpeckleTopSolid.UI.csproj", "{155758C6-E385-4284-8700-3C3F5A52090C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConverterTopSolid", "UI\ConverterTopSolid\ConverterTopSolid.csproj", "{B806165A-12EF-4E99-9D0E-717CFBE93756}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,6 +33,14 @@ Global
 		{155758C6-E385-4284-8700-3C3F5A52090C}.Release|Any CPU.Build.0 = Release|x64
 		{155758C6-E385-4284-8700-3C3F5A52090C}.Release|x64.ActiveCfg = Release|x64
 		{155758C6-E385-4284-8700-3C3F5A52090C}.Release|x64.Build.0 = Release|x64
+		{B806165A-12EF-4E99-9D0E-717CFBE93756}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B806165A-12EF-4E99-9D0E-717CFBE93756}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B806165A-12EF-4E99-9D0E-717CFBE93756}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B806165A-12EF-4E99-9D0E-717CFBE93756}.Debug|x64.Build.0 = Debug|Any CPU
+		{B806165A-12EF-4E99-9D0E-717CFBE93756}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B806165A-12EF-4E99-9D0E-717CFBE93756}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B806165A-12EF-4E99-9D0E-717CFBE93756}.Release|x64.ActiveCfg = Release|Any CPU
+		{B806165A-12EF-4E99-9D0E-717CFBE93756}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
