+++ UI/ConverterTopSolid/ConverterTopSolid.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConverterTopSolid", "ConverterTopSolid.csproj", "{2B0812F6-62E6-491A-949D-F42B769B7502}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2B0812F6-62E6-491A-949D-F42B769B7502}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B0812F6-62E6-491A-949D-F42B769B7502}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2B0812F6-62E6-491A-949D-F42B769B7502}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2B0812F6-62E6-491A-949D-F42B769B7502}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
